pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "device-scanner"
version = "0.1.0"
description = "A device scanner application"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Shmaya Peskin", email = "shmayaky@gmail.com" }]
dependencies = ["xmltodict", "sqlalchemy", "netifaces", "python-dotenv"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff", "pylint"]

[tool.setuptools]
packages = ["device_scanner"]
